[NVIDIA] Add cuda_perf_counts.hpp
